project/requirements.txt
asgiref==3.7.2
autopep8==2.0.2
certifi==2023.7.22
Django==4.2.4
django-elasticsearch-dsl==7.3
django-silk==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
elastic-transport==8.4.0
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
Faker==19.3.0
gprof2dot==2022.7.29
pycodestyle==2.11.0
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
urllib3==1.26.16
